+++ src/iomarket.cpp
@@ -22,7 +22,7 @@ MarketOfferList IOMarket::getActiveOffers(MarketAction_t action, uint16_t itemId
 
 	DBResult_ptr result = Database::getInstance().storeQuery(fmt::format(
 	    "SELECT `id`, `amount`, `price`, `created`, `anonymous`, (SELECT `name` FROM `players` WHERE `id` = `player_id`) AS `player_name` FROM `market_offers` WHERE `sale` = {:d} AND `itemtype` = {:d}",
-	    action, itemId));
+	    static_cast<uint8_t>(action), itemId));
 	if (!result) {
 		return offerList;
 	}
@@ -53,7 +53,7 @@ MarketOfferList IOMarket::getOwnOffers(MarketAction_t action, uint32_t playerId)
 
 	DBResult_ptr result = Database::getInstance().storeQuery(fmt::format(
 	    "SELECT `id`, `amount`, `price`, `created`, `itemtype` FROM `market_offers` WHERE `player_id` = {:d} AND `sale` = {:d}",
-	    playerId, action));
+	    playerId, static_cast<uint8_t>(action)));
 	if (!result) {
 		return offerList;
 	}
@@ -76,7 +76,7 @@ HistoryMarketOfferList IOMarket::getOwnHistory(MarketAction_t action, uint32_t p
 
 	DBResult_ptr result = Database::getInstance().storeQuery(fmt::format(
 	    "SELECT `itemtype`, `amount`, `price`, `expires_at`, `state` FROM `market_history` WHERE `player_id` = {:d} AND `sale` = {:d}",
-	    playerId, action));
+	    playerId, static_cast<uint8_t>(action)));
 	if (!result) {
 		return offerList;
 	}
@@ -242,7 +242,7 @@ void IOMarket::createOffer(uint32_t playerId, MarketAction_t action, uint32_t it
 {
 	Database::getInstance().executeQuery(fmt::format(
 	    "INSERT INTO `market_offers` (`player_id`, `sale`, `itemtype`, `amount`, `price`, `created`, `anonymous`) VALUES ({:d}, {:d}, {:d}, {:d}, {:d}, {:d}, {:d})",
-	    playerId, action, itemId, amount, price, time(nullptr), anonymous));
+	    playerId, static_cast<uint8_t>(action), itemId, amount, price, time(nullptr), anonymous));
 }
 
 void IOMarket::acceptOffer(uint32_t offerId, uint16_t amount)
@@ -261,7 +261,8 @@ void IOMarket::appendHistory(uint32_t playerId, MarketAction_t type, uint16_t it
 {
 	g_databaseTasks.addTask(fmt::format(
 	    "INSERT INTO `market_history` (`player_id`, `sale`, `itemtype`, `amount`, `price`, `expires_at`, `inserted`, `state`) VALUES ({:d}, {:d}, {:d}, {:d}, {:d}, {:d}, {:d}, {:d})",
-	    playerId, type, itemId, amount, price, timestamp, time(nullptr), state));
+	    playerId, static_cast<uint8_t>(type), itemId, amount, price, timestamp, time(nullptr),
+	    static_cast<uint8_t>(state)));
 }
 
 bool IOMarket::moveOfferToHistory(uint32_t offerId, MarketOfferState_t state)
@@ -292,7 +293,7 @@ void IOMarket::updateStatistics()
 {
 	DBResult_ptr result = Database::getInstance().storeQuery(fmt::format(
 	    "SELECT `sale` AS `sale`, `itemtype` AS `itemtype`, COUNT(`price`) AS `num`, MIN(`price`) AS `min`, MAX(`price`) AS `max`, SUM(`price`) AS `sum` FROM `market_history` WHERE `state` = {:d} GROUP BY `itemtype`, `sale`",
-	    OFFERSTATE_ACCEPTED));
+	    static_cast<uint8_t>(OFFERSTATE_ACCEPTED)));
 	if (!result) {
 		return;
 	}
